added basic transformer auto complete, performed some code refactoring
