+++ Raven.Studio.Html5/App/views/editIndex.html
@@ -77,7 +77,7 @@
                 <div data-bind="foreach: maps" class="col-md-11">
                     <div class="row">
                         <div class="col-md-11">
-                            <pre id="indexEditor" class="form-control map" data-bind="aceEditor: { code: $data, getFocus: true, lang:'ace/mode/ravenMapLinq', completer:$root.indexMapCompleter, typeName:'map', containigViewModel:$root}"></pre>
+                            <pre id="indexEditor" class="form-control map" data-bind="aceEditor: { code: $data, getFocus: true, lang:'ace/mode/ravenMapLinq', completer:$root.indexAutoCompleter.indexMapCompleter, typeName:'map', containigViewModel:$root.indexAutoCompleter}"></pre>
                         </div>
                         <button type="button" class="close pull-left"  data-bind=" visible: $index() !== 0 || $root.hasMultipleMaps(), click: $root.removeMap.bind($root, $index())" aria-hidden="true" title="Remove this map from the index"><span class="glyphicon glyphicon-remove remove-icon"></span></button>
                     </div>
@@ -92,7 +92,7 @@
                 <div class="col-md-11">
                     <div class="row">
                         <div class="col-md-11">
-                            <pre class="form-control map" data-bind="aceEditor: { code: reduce, lang:'ace/mode/ravenReduceLinq',completer:$root.indexReduceCompleter, typeName:'reduce', containigViewModel:$root }"></pre>
+                            <pre class="form-control map" data-bind="aceEditor: { code: reduce, lang:'ace/mode/ravenReduceLinq',completer:$root.indexAutoCompleter.indexReduceCompleter, typeName:'reduce', containigViewModel:$root.indexAutoCompleter }"></pre>
                         </div>
                         <button type="button" class="close pull-left"  data-bind="click: $root.removeReduce.bind($root)" aria-hidden="true" title="Remove the Reduce function from the index"><span class="glyphicon glyphicon-remove remove-icon"></span></button>
                     </div>
@@ -106,7 +106,7 @@
                 <div class="col-md-11">
                     <div class="row">
                         <div class="col-md-11">
-                            <pre class="form-control map" data-bind="aceEditor: { code: transformResults, lang:'ace/mode/ravenLinq' }"></pre>
+                            <pre class="form-control map" data-bind="aceEditor: { code: transformResults, lang:'ace/mode/ravenLinq',lang:'ace/mode/ravenMapLinq', completer:$root.indexAutoCompleter.indexReduceCompleter, typeName:'transform', containigViewModel:$root.indexAutoCompleter }"></pre>
                         </div>
                          <button type="button" class="close pull-left"  data-bind="click: $root.removeTransform.bind($root)" aria-hidden="true" title="Remove Transform Results from the index"><span class="glyphicon glyphicon-remove remove-icon"></span></button>
                     </div>
