+++ Raven.Studio.Html5/Scripts/ace/mode/ravenReduceLinq_highlight_rules.js
@@ -26,9 +26,7 @@
                 },
                 {
                     token: "data.prefix",
-                    regex: "\\w+\\.",
-                    next: "data.prefix"
-
+                    regex: "\\w+\\."
                 },
                 DocCommentHighlightRules.getStartRule("doc-start"),
                 {
